feature(cardsFromSequence): Initial implementation to show cards in a specific order, no matter what choices the player makes.
Since the nodes in a CardSequence are CardTrees, each node can be a branching narrative.
By organizing CardTrees in a CardSequence, all CardLeafs (endings) within the CardTree branching narrative will come back to the same spot.

This allows us to create scenarios with multiple subplots (CardTrees) where the player can get different endings, but always continue with the next node in the CardSequence.

Should be pretty fun to play with!